requirements.txt
Django==1.6.1
PIL==1.1.7
distribute==0.6.27
django-tastypie==0.11.0
pysqlite==2.6.3
python-dateutil==2.2
python-mimeparse==0.1.4
six==1.4.1
wsgiref==0.1.2
